Move the drawBOunds job to DrawTasks
