+++ libraries/render-utils/src/RenderForwardTask.cpp
@@ -24,8 +24,6 @@
 
 #include <gpu/StandardShaderLib.h>
 
-#include <render/drawItemBounds_vert.h>
-#include <render/drawItemBounds_frag.h>
 #include "nop_frag.h"
 
 using namespace render;
@@ -182,57 +180,4 @@ void DrawBackground::run(const SceneContextPointer& sceneContext, const RenderCo
     args->_batch = nullptr;
 }
 
-const gpu::PipelinePointer DrawBounds::getPipeline() {
-    if (!_boundsPipeline) {
-        auto vs = gpu::Shader::createVertex(std::string(drawItemBounds_vert));
-        auto ps = gpu::Shader::createPixel(std::string(drawItemBounds_frag));
-        gpu::ShaderPointer program = gpu::Shader::createProgram(vs, ps);
 
-        gpu::Shader::BindingSet slotBindings;
-        gpu::Shader::makeProgram(*program, slotBindings);
-
-        _cornerLocation = program->getUniforms().findLocation("inBoundPos");
-        _scaleLocation = program->getUniforms().findLocation("inBoundDim");
-
-        auto state = std::make_shared<gpu::State>();
-        state->setDepthTest(true, false, gpu::LESS_EQUAL);
-        state->setBlendFunction(true,
-            gpu::State::SRC_ALPHA, gpu::State::BLEND_OP_ADD, gpu::State::INV_SRC_ALPHA,
-            gpu::State::DEST_ALPHA, gpu::State::BLEND_OP_ADD, gpu::State::ZERO);
-
-        _boundsPipeline = gpu::Pipeline::create(program, state);
-    }
-    return _boundsPipeline;
-}
-
-void DrawBounds::run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext,
-        const Inputs& items) {
-    RenderArgs* args = renderContext->args;
-
-    gpu::doInBatch(args->_context, [&](gpu::Batch& batch) {
-        args->_batch = &batch;
-
-        // Setup projection
-        glm::mat4 projMat;
-        Transform viewMat;
-        args->getViewFrustum().evalProjectionMatrix(projMat);
-        args->getViewFrustum().evalViewTransform(viewMat);
-        batch.setProjectionTransform(projMat);
-        batch.setViewTransform(viewMat);
-        batch.setModelTransform(Transform());
-
-        // Bind program
-        batch.setPipeline(getPipeline());
-        assert(_cornerLocation >= 0);
-        assert(_scaleLocation >= 0);
-
-        // Render bounds
-        for (const auto& item : items) {
-            batch._glUniform3fv(_cornerLocation, 1, (const float*)(&item.bound.getCorner()));
-            batch._glUniform3fv(_scaleLocation, 1, (const float*)(&item.bound.getScale()));
-
-            static const int NUM_VERTICES_PER_CUBE = 24;
-            batch.draw(gpu::LINES, NUM_VERTICES_PER_CUBE, 0);
-        }
-    });
-}
