+++ libraries/render-utils/src/RenderForwardTask.h
@@ -68,25 +68,4 @@ public:
             const Inputs& background);
 };
 
-class DrawBounds {
-public:
-    class Config : public render::JobConfig {
-    public:
-        Config() : JobConfig(false) {}
-    };
-
-    using Inputs = render::ItemBounds;
-    using JobModel = render::Job::ModelI<DrawBounds, Inputs, Config>;
-
-    void configure(const Config& configuration) {}
-    void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext,
-            const Inputs& items);
-
-private:
-    const gpu::PipelinePointer getPipeline();
-    gpu::PipelinePointer _boundsPipeline;
-    int _cornerLocation { -1 };
-    int _scaleLocation { -1 };
-};
-
 #endif // hifi_RenderForwardTask_h
